# api/openapi.yaml
# yaml-language-server: $schema=https://spec.openapis.org/oas/3.0/schema/2024-10-18
openapi: 3.0.4
info:
  title: Sonimulus API
  version: 0.1.0
servers:
  - url: https://localhost:3000
paths:
  /auth:
    put:
      summary: Perfoms an account registration request
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - name
                - email
                - password
              additionalProperties: false
      responses:
        "201":
          description: Account registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/authResponse"
        "409":
          description: Email already in use
    post:
      summary: Perfoms a login request
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
              additionalProperties: false
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/authResponse"
        "401":
          description: Login failed
  /users/{name}:
    get:
      summary: Gets a user by their name
      operationId: getUserByName
      security:
        - jwtAuth: []
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
          description: String name of the user to get
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    authResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        refresh_token:
          type: string
        user:
          $ref: "#/components/schemas/user"
      required:
        - access_token
        - token_type
        - expires_in
        - refresh_token
        - user
      additionalProperties: false
    user:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
      required:
        - id
        - name
        - email
      additionalProperties: false